requirements.txt
pycryptodome
websockets
docopt